requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.9
Django==5.1.6
django-ckeditor==6.7.2
django-js-asset==3.0.1
djangorestframework==3.15.2
filelock==3.16.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.1.0
platformdirs==4.3.6
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.27.1
Werkzeug==3.0.6
